refactor: `ConnectionController`로 통합 및 환경 변수 추가
