+++ backend/src/main/resources/application-prod.yaml
@@ -10,7 +10,9 @@ spring:
       hibernate:
         format_sql: true
     show-sql: true
-kakao_client_secret: ${KAKAO_CLIENT_SECRET}
 kakao_client_id: ${KAKAO_CLIENT_ID}
+kakao_client_secret: ${KAKAO_CLIENT_SECRET}
 tistory_client_id: ${TISTORY_CLIENT_ID}
 tistory_client_secret: ${TISTORY_CLIENT_SECRET}
+notion_client_id: ${NOTION_CLIENT_ID}
+notion_client_secret: ${NOTION_CLIENT_SECRET}
