+++ backend/src/main/resources/application.yaml
@@ -15,7 +15,9 @@ spring:
   h2:
     console:
       enabled: true
-kakao_client_secret: ${KAKAO_CLIENT_SECRET}
 kakao_client_id: ${KAKAO_CLIENT_ID}
+kakao_client_secret: ${KAKAO_CLIENT_SECRET}
 tistory_client_id: ${TISTORY_CLIENT_ID}
 tistory_client_secret: ${TISTORY_CLIENT_SECRET}
+notion_client_id: ${NOTION_CLIENT_ID}
+notion_client_secret: ${NOTION_CLIENT_SECRET}
